requirements-global.txt
# base requirements
numpy==2.2.6
opencv-python==4.11.0.86
pillow==11.3.0
imagesize==1.4.1 #for concept statistics
tqdm==4.67.1
PyYAML==6.0.2
huggingface-hub==0.34.4
scipy==1.15.3
matplotlib==3.10.3
av==14.4.0
yt-dlp  #no pinned version, frequently updated for compatibility with sites
scenedetect==0.6.6

# pytorch
accelerate==1.7.0
safetensors==0.5.3
tensorboard==2.19.0
pytorch-lightning==2.5.1.post0

# diffusion models
#Note: check whether Qwen bugs in diffusers have been fixed before upgrading diffusers (see BaseQwenSetup):
-e git+https://github.com/huggingface/diffusers.git@9b721db#egg=diffusers

transformers==4.52.4
sentencepiece==0.2.0 # transitive dependency of transformers for tokenizer loading
omegaconf==2.3.0 # needed to load stable diffusion from single ckpt files
invisible-watermark==0.2.0 # needed for the SDXL pipeline

# model conversion
-e git+https://github.com/Open-Model-Initiative/OMI-Model-Standards.git@f14b1da#egg=omi_model_standards

# other models
pooch==1.8.2
open-clip-torch==2.32.0

# data loader
-e git+https://github.com/Nerogar/mgds.git@50a2394#egg=mgds

# optimizers
dadaptation==3.2 # dadaptation optimizers
lion-pytorch==0.2.3 # lion optimizer
prodigyopt==1.1.2 # prodigy optimizer
schedulefree==1.4.1 # schedule-free optimizers
pytorch_optimizer==3.6.0 # pytorch optimizers
prodigy-plus-schedule-free==2.0.0 # Prodigy plus optimizer
adv_optm==1.0.6 # advanced optimizers

# Profiling
scalene==1.5.51

# ui
customtkinter==5.2.2

# cloud
runpod==1.7.10
fabric==3.2.2

# debug
psutil==7.0.0
requests==2.32.3
deepdiff==8.6.1 # output easy to read diff for troublshooting
loguru==0.7.2
peft==0.13.1
einops==0.8.0
flash-attn==2.6.3; platform_system != "Windows"  # Optional fused attention kernels (skip on Windows where no wheels)
decord==0.6.0
